extract views
