--- src/colonization-costs.json
@@ -1,213 +1,0 @@
-{
-  "coriolis": {
-    "liquidoxygen": 3781,
-    "water": 1609,
-    "ceramiccomposites": 1207,
-    "cmmcomposite": 11261,
-    "insulatingmembrane": 644,
-    "polymers": 1046,
-    "semiconductors": 161,
-    "superconductors": 282,
-    "aluminium": 10055,
-    "copper": 644,
-    "steel": 14076,
-    "titanium": 8205,
-    "computercomponents": 145,
-    "medicaldiagnosticequipment": 25,
-    "foodcartridges": 242,
-    "fruitandvegetables": 145,
-    "nonlethalweapons": 25,
-    "powergenerators": 65,
-    "waterpurifiers": 105
-  },
-  "asteroid": {
-    "liquidoxygen": 3781,
-    "water": 1609,
-    "ceramiccomposites": 1207,
-    "cmmcomposite": 11261,
-    "insulatingmembrane": 644,
-    "polymers": 1046,
-    "semiconductors": 161,
-    "superconductors": 282,
-    "aluminium": 10055,
-    "copper": 644,
-    "steel": 14076,
-    "titanium": 8205,
-    "computercomponents": 145,
-    "medicaldiagnosticequipment": 25,
-    "foodcartridges": 242,
-    "fruitandvegetables": 145,
-    "nonlethalweapons": 25,
-    "powergenerators": 65,
-    "waterpurifiers": 105
-  },
-  "ocellus": {
-    "liquidoxygen": 15124,
-    "water": 6436,
-    "ceramiccomposites": 4828,
-    "cmmcomposite": 45044,
-    "insulatingmembrane": 1288,
-    "polymers": 2092,
-    "semiconductors": 322,
-    "superconductors": 564,
-    "aluminium": 40220,
-    "copper": 2576,
-    "steel": 56304,
-    "titanium": 32820,
-    "computercomponents": 290,
-    "medicaldiagnosticequipment": 50,
-    "foodcartridges": 484,
-    "fruitandvegetables": 290,
-    "nonlethalweapons": 50,
-    "powergenerators": 130,
-    "waterpurifiers": 210
-  },
-  "orbis": {
-    "liquidoxygen": 15124,
-    "water": 6436,
-    "ceramiccomposites": 4828,
-    "cmmcomposite": 45044,
-    "insulatingmembrane": 1288,
-    "polymers": 2092,
-    "semiconductors": 322,
-    "superconductors": 564,
-    "aluminium": 40220,
-    "copper": 2576,
-    "steel": 56304,
-    "titanium": 32820,
-    "computercomponents": 290,
-    "medicaldiagnosticequipment": 50,
-    "foodcartridges": 484,
-    "fruitandvegetables": 290,
-    "nonlethalweapons": 50,
-    "powergenerators": 130,
-    "waterpurifiers": 210
-  },
-  "outpost-commercial": {
-    "liquidoxygen": 1553,
-    "water": 621,
-    "ceramiccomposites": 497,
-    "cmmcomposite": 3912,
-    "insulatingmembrane": 311,
-    "polymers": 497,
-    "semiconductors": 56,
-    "superconductors": 100,
-    "aluminium": 515,
-    "copper": 218,
-    "steel": 5588,
-    "titanium": 4843,
-    "computercomponents": 50,
-    "medicaldiagnosticequipment": 13,
-    "foodcartridges": 94,
-    "fruitandvegetables": 50,
-    "nonlethalweapons": 13,
-    "powergenerators": 19,
-    "waterpurifiers": 38
-  },
-  "outpost-science": {
-    "liquidoxygen": 1553,
-    "water": 621,
-    "ceramiccomposites": 497,
-    "cmmcomposite": 3912,
-    "insulatingmembrane": 311,
-    "polymers": 497,
-    "semiconductors": 56,
-    "superconductors": 100,
-    "aluminium": 515,
-    "copper": 218,
-    "steel": 5588,
-    "titanium": 4843,
-    "computercomponents": 50,
-    "medicaldiagnosticequipment": 13,
-    "foodcartridges": 94,
-    "fruitandvegetables": 50,
-    "nonlethalweapons": 13,
-    "powergenerators": 19,
-    "waterpurifiers": 38
-  },
-  "outpost-industrial": {
-    "liquidoxygen": 1553,
-    "water": 621,
-    "ceramiccomposites": 497,
-    "cmmcomposite": 3912,
-    "insulatingmembrane": 311,
-    "polymers": 497,
-    "semiconductors": 56,
-    "superconductors": 100,
-    "aluminium": 515,
-    "copper": 218,
-    "steel": 5588,
-    "titanium": 4843,
-    "computercomponents": 50,
-    "medicaldiagnosticequipment": 13,
-    "foodcartridges": 94,
-    "fruitandvegetables": 50,
-    "nonlethalweapons": 13,
-    "powergenerators": 19,
-    "waterpurifiers": 38
-  },
-  "outpost-criminal": {
-    "liquidoxygen": 1553,
-    "water": 621,
-    "ceramiccomposites": 497,
-    "cmmcomposite": 3912,
-    "insulatingmembrane": 311,
-    "polymers": 497,
-    "semiconductors": 56,
-    "superconductors": 100,
-    "aluminium": 515,
-    "copper": 218,
-    "steel": 5588,
-    "titanium": 4843,
-    "computercomponents": 50,
-    "medicaldiagnosticequipment": 13,
-    "foodcartridges": 94,
-    "fruitandvegetables": 50,
-    "nonlethalweapons": 13,
-    "powergenerators": 19,
-    "waterpurifiers": 38
-  },
-  "outpost-civilian": {
-    "liquidoxygen": 1553,
-    "water": 621,
-    "ceramiccomposites": 497,
-    "cmmcomposite": 3912,
-    "insulatingmembrane": 311,
-    "polymers": 497,
-    "semiconductors": 56,
-    "superconductors": 100,
-    "aluminium": 515,
-    "copper": 218,
-    "steel": 5588,
-    "titanium": 4843,
-    "computercomponents": 50,
-    "medicaldiagnosticequipment": 13,
-    "foodcartridges": 94,
-    "fruitandvegetables": 50,
-    "nonlethalweapons": 13,
-    "powergenerators": 19,
-    "waterpurifiers": 38
-  },
-  "outpost-military": {
-    "liquidoxygen": 1553,
-    "water": 621,
-    "ceramiccomposites": 497,
-    "cmmcomposite": 3912,
-    "insulatingmembrane": 311,
-    "polymers": 497,
-    "semiconductors": 56,
-    "superconductors": 100,
-    "aluminium": 515,
-    "copper": 218,
-    "steel": 5588,
-    "titanium": 4843,
-    "computercomponents": 50,
-    "medicaldiagnosticequipment": 13,
-    "foodcartridges": 94,
-    "fruitandvegetables": 50,
-    "nonlethalweapons": 13,
-    "powergenerators": 19,
-    "waterpurifiers": 38
-  }
-
-}
