+++ src/views/index.ts
@@ -1,0 +1,5 @@
+export { About } from "./About";
+export { Commander } from "./Commander";
+export { Home } from "./Home";
+export { ProjectSearch } from "./ProjectSearch";
+export { ProjectView } from "./ProjectView";
